[2/N] Refactor FR script - add a loader module
[ghstack-poisoned]
